Divided project in subprojects and added different directories for better organisation
