+++ Back-end/Entities/offer.entity.ts
