+++ Back-end/Entities/slide.entity.ts
