+++ Back-end/Modules/slide.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { SlideService } from '../Services/slide.service';
+import { SlideController } from '../Controller/slide.controller';
+import { Slide } from '../Entities/slide.entity';
+
+@Module({
+    imports: [TypeOrmModule.forFeature([Slide])],
+    providers: [SlideService],
+    controllers: [SlideController],
+})
+export class ThematicModule {}
