+++ Back-end/Services/user.service.ts
@@ -1,0 +1,28 @@
+import { Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+import { User } from '../Entities/user.entity';
+
+@Injectable()
+export class UsersService {
+    constructor(
+        @InjectRepository(User)
+        private usersRepository: Repository<User>,
+    ) {}
+
+    findAll(): Promise<User[]> {
+        return this.usersRepository.find();
+    }
+
+    async createUser(user : User){
+        const newUser = await this.usersRepository.create(user);
+    }
+
+    findOne(id: string): Promise<User> {
+        return this.usersRepository.findOne(id);
+    }
+
+    async remove(id: string): Promise<void> {
+        await this.usersRepository.delete(id);
+    }
+}
