+++ Back-end/src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
+import { AppController } from '../Controller/app.controller';
 import { AppService } from './app.service';
 
 @Module({
