+++ Back-end/src/app.service.ts
@@ -1,0 +1,9 @@
+
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class AppService {
+    getHello(): string {
+        return 'Hello World!';
+    }
+}
