+++ Back-end/src/main.ts
@@ -1,0 +1,9 @@
+
+import { NestFactory } from '@nestjs/core';
+import { DatabaseModule } from './database.module';
+
+async function bootstrap() {
+    const app = await NestFactory.create(DatabaseModule);
+    await app.listen(3000);
+}
+bootstrap();
