--- src/Entities/media/media.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { MediaService } from './media.service';
-import { MediaController } from './media.controller';
-import { Media } from './media.entity';
-
-@Module({
-    imports: [TypeOrmModule.forFeature([Media])],
-    providers: [MediaService],
-    controllers: [MediaController],
-})
-export class MediaModule {}
